requirements.txt
PyJWT>=2.7.0
requests>=2.31.0
cryptography>=41.0.1
python-dotenv>=1.0.0
